requirements.txt
appdirs==1.4.0
astroid==1.4.9
backports.functools-lru-cache==1.3
configparser==3.5.0
dj-database-url==0.4.2
Django==1.10.5
django-filter==1.0.1
django-taggit==0.22.0
djangorestframework==3.5.3
gunicorn==19.6.0
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
mccabe==0.6.1
packaging==16.8
psycopg2==2.6.2
pylint==1.6.5
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
wrapt==1.10.8
